backend/requirements.txt
Flask
pymongo
flask-bcrypt
Flask-PyMongo
Flask-JWT-Extended
python-dotenv
Flask-CORS